Leaf Disease Classification
CNNs are a sort of neural network that performs particularly well in image categorization tasks. They can automatically learn and extract information from images, which they subsequently use to classify into several categories.
The purpose of leaf image classification for vegetables is to train a CNN to recognise and categorise various types of vegetable leaves. This can be accomplished by feeding the CNN with a huge collection of labelled leaf images, each of which corresponds to a specific vegetable species.
The CNN will then learn to extract elements from the photos that are useful in identifying between the various vegetable species. For example, it may learn to recognise the form, colour, and texture of various leaves.